+++ src/inmanta/export.py
@@ -480,6 +480,14 @@ class Exporter(object):
 
         return hash_id
 
+    def get_environment_id(self) -> str:
+        env = str(cfg_env.get())
+
+        if env is None:
+            raise Exception("The environment of the model should be configured in config>environment")
+
+        return env
+
 
 class dependency_manager(object):  # noqa: N801
     """